requirements.txt
Django
django-filter
mysqlclient
pillow
django-cleanup
python-dotenv
google-generativeai